Corrección de modelos y orden de proyecto
