+++ trello_backend/apps/tasks/models/priority.py
@@ -4,7 +4,7 @@ from apps.tasks.models.auditor import Auditor
 
 class Priority(Auditor):
     
-    name = models.CharField(max_length=100, null=True, blank=True)
+    name = models.CharField(max_length=100, null=False)
 
     def __str__(self) -> str:
         return "{}".format(self.name)
