+++ trello_backend/apps/tasks/views/task_view.py
@@ -1,7 +1,7 @@
 from rest_framework import generics
 
 from apps.tasks.models import Task
-from apps.tasks.api.serializers.task_serializer import TaskSerializer, TaskListSerializer
+from apps.tasks.serializers.task_serializer import TaskSerializer, TaskListSerializer
 from apps.tasks.filters import TaskFilter
