+++ trello_backend/requirements.txt
@@ -1,10 +1,16 @@
 asgiref==3.6.0
-backports.zoneinfo==0.2.1
+class-registry==2.1.2
 Django==4.1.7
-django-cors-headers==3.13.0
+django-cors-headers==3.14.0
 django-filter==22.1
 djangorestframework==3.14.0
-gunicorn==20.1.0
-psycopg2-binary==2.9.5
+filters-django==1.0.5
+pip==22.0.4
+psycopg2==2.9.5
+python-dateutil==2.8.2
 pytz==2022.7.1
+regex==2022.10.31
+setuptools==58.1.0
+six==1.16.0
 sqlparse==0.4.3
+tzdata==2022.7
