+++ trello_backend/trello/urls.py
@@ -3,5 +3,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('tasks/', include('apps.tasks.api.urls')),
+    path('tasks/', include('apps.tasks.urls')),
 ]
